feat: add status page with backend API integration for HP calculation
