+++ frontend/app/layout.tsx
@@ -1,33 +1,48 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+import Link from "next/link";
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+export const metadata = {
+  title: "Zero-Lab",
+  description: "Pokémon Battle Assist AI",
 };
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
-    <html
-      lang="en"
-      className={`${geistSans.variable} ${geistMono.variable} h-full antialiased`}
-    >
-      <body className="min-h-full flex flex-col">{children}</body>
+    <html lang="ja">
+      <body className="bg-gray-50 min-h-screen flex flex-col font-sans">
+        {/* ▼▼▼ 全ページ共通のナビゲーションバー ▼▼▼ */}
+        <header className="bg-gray-900 text-white shadow-md">
+          <div className="max-w-5xl mx-auto px-4 py-4 flex justify-between items-center">
+            {/* 左側のロゴ（クリックでトップページへ） */}
+            <Link href="/" className="text-xl font-extrabold tracking-widest hover:text-blue-400 transition">
+              ZERO-LAB
+            </Link>
+
+            {/* 右側のメニューリンク群 */}
+            <nav className="flex gap-6 text-sm font-medium">
+              <Link href="/status" className="hover:text-blue-400 transition">
+                ステータス計算
+              </Link>
+              <Link href="/damage" className="hover:text-blue-400 transition text-gray-400">
+                ダメージ計算(準備中)
+              </Link>
+              <Link href="/pokedex" className="hover:text-blue-400 transition text-gray-400">
+                ポケモン図鑑(準備中)
+              </Link>
+            </nav>
+          </div>
+        </header>
+        {/* ▲▲▲ ナビゲーションバー ここまで ▲▲▲ */}
+
+        {/* 各ページの中身（page.tsx）がこの {children} の部分に差し込まれます */}
+        <div className="flex-1 max-w-5xl mx-auto w-full p-4 md:p-8">
+          {children}
+        </div>
+      </body>
     </html>
   );
 }
